Move EnumerableTests to Collections namespace
